Corrections for Qt 5.1
git-svn-id: http://b035a0a.online-server.cloud/svn/AcfSln@980 3363cc13-2b78-0410-9281-ee1caeb4462d

+++ Include/imeas/TDiscreteDataSequence.h
@@ -120,8 +120,8 @@ bool TDiscreteDataSequence<Element>::CreateDiscreteSequence(
 	m_allocatedElementsCount = 0;
 	m_samplesCount = samplesCount;
 	m_channelsCount = channelsCount;
-	m_sampleDiff = (sampleDiff != 0)? sampleDiff: channelsCount * sizeof(Element);
-	m_channelDiff = (channelDiff != 0)? channelDiff: sizeof(Element);
+	m_sampleDiff = (sampleDiff != 0) ? sampleDiff : int(channelsCount * sizeof(Element));
+	m_channelDiff = (channelDiff != 0) ? channelDiff: sizeof(Element);
 
 	return true;
 }
